- BugFix: mixing of long and int lead to overflow on 64bit platform
[r24329]

+++ ir/be/ia32/ia32_transform.c
@@ -2568,8 +2568,8 @@ static ir_node *create_Switch(ir_node *node)
 	ir_node  *block      = be_transform_node(get_nodes_block(node));
 	ir_node  *sel        = get_Cond_selector(node);
 	ir_node  *new_sel    = be_transform_node(sel);
-	int       switch_min = INT_MAX;
-	int       switch_max = INT_MIN;
+	long      switch_min = LONG_MAX;
+	long      switch_max = LONG_MIN;
 	long      default_pn = get_Cond_defaultProj(node);
 	ir_node  *new_node;
 	const ir_edge_t *edge;
@@ -2589,7 +2589,7 @@ static ir_node *create_Switch(ir_node *node)
 			switch_max = pn;
 	}
 
-	if ((unsigned) (switch_max - switch_min) > 256000) {
+	if ((unsigned long) (switch_max - switch_min) > 256000) {
 		panic("Size of switch %+F bigger than 256000", node);
 	}
